Computing per-plate average for raw absorbance of the standard blank

This only makes sense for plates that have contain 2 or more standard curves for the same standard solution

Usage

std_blank_average(std_blank_data)

Arguments

std_blank_data

A tibble containing only rows that received standard blanks, i.e., row A (top_down_pipetting) or H (bottom_up_pipetting). It is normally generated by extract_std_blank() from which one element can be chosen and, if necessary, modified, based on the examination of "untrusted wells". Thus, use as input (a modified version of) extract_std_blank(data)$all or extract_std_blank(data)$trusted

Value

A tibble with one row per plate, and columns dataset, plate_id, blank_avg, blank_sdev, blank_coeff_var_percent

Examples

tidy_plates
#> # A tibble: 480 × 8
#>    row   column well_id unique_well_id dataset plate_id map      abs  
#>    <chr> <chr>  <chr>   <chr>          <chr>   <chr>    <chr>    <chr>
#>  1 A     1      A1      A1_NO3_1F1     Nmin    NO3_1F1  Std      0.092
#>  2 A     1      A1      A1_NO3_1F2     Nmin    NO3_1F2  Std      0.091
#>  3 A     1      A1      A1_NO3_1F3     Nmin    NO3_1F3  Std      0.110
#>  4 A     1      A1      A1_NO3_1F4     Nmin    NO3_1F4  Std      0.092
#>  5 A     1      A1      A1_NO3_1F5     Nmin    NO3_1F5  Std      0.113
#>  6 A     2      A2      A2_NO3_1F1     Nmin    NO3_1F1  81_t1_z2 0.114
#>  7 A     2      A2      A2_NO3_1F2     Nmin    NO3_1F2  97_t1_z1 0.107
#>  8 A     2      A2      A2_NO3_1F3     Nmin    NO3_1F3  89_t1_z3 0.095
#>  9 A     2      A2      A2_NO3_1F4     Nmin    NO3_1F4  81_t1_z1 0.118
#> 10 A     2      A2      A2_NO3_1F5     Nmin    NO3_1F5  Std_3_t1 0.167
#> # ℹ 470 more rows
std_blank <- tidy_plates |> extract_std_blank(std_def = "Std")
std_blank$trusted |> std_blank_average()
#> # A tibble: 5 × 5
#>   dataset plate_id blank_avg blank_sdev blank_coeff_var_percent
#>   <chr>   <chr>        <dbl>      <dbl>                   <dbl>
#> 1 Nmin    NO3_1F1     0.0915   0.000707                   0.773
#> 2 Nmin    NO3_1F2     0.0905   0.000707                   0.781
#> 3 Nmin    NO3_1F4     0.0915   0.000707                   0.773
#> 4 Nmin    NO3_1F3     0.09    NA                         NA    
#> 5 Nmin    NO3_1F5     0.092   NA                         NA    
# Don't be disturbed by the NA values: it is not possible to compute a
# standard deviation of coefficient of variation from only one data entry